+++ iOSswiftUIa11yTechniques/iOSswiftUIa11yTechniques/ContentView.swift
@@ -196,6 +196,7 @@ struct ContentView: View {
         case "reduce motion": ReduceMotionView()
         case "reduce transparency": ReduceTransparencyView()
         case "smart invert": SmartInvertView()
+        case "a11y-check": A11yCheckView()
         default: InformativeView()
         }
     }