+++ iOSswiftUIa11yTechniques/iOSswiftUIa11yTechniques/Techniques.swift
@@ -85,5 +85,6 @@ let techniques: [Techniques] = [
     Techniques(name: "Videos"),
     Techniques(name: "Touch Target Size"),
     Techniques(name: "VoiceOver Announcement Delay"),
-    Techniques(name: "VoiceOver Pronunciation")
+    Techniques(name: "VoiceOver Pronunciation"),
+    Techniques(name: "a11y-check")
 ]